refactor : 오류코드 전부수정
