+++ src/Router.jsx
@@ -7,8 +7,8 @@ import SearchPage from "./pages/Search";
 import NotFoundScene from "./pages/NotFoundScene";
 import useUserData from "./api/useUserData";
 import { useEffect } from "react";
-import BoardEditPost from "./components/Board/BoardEditPost";
-import ThreeBoardEditPost from "./components/Board/ThreePower/ThreeBoardEditPost";
+import BoardEditPost from "./pages/BoardEditPost";
+import ThreeBoardEditPost from "./pages/ThreeBoardEditPost";
 import Board from "./pages/Board";
 import BoardPost from "./pages/BoardPost";
 import BoardWrite from "./pages/BoardWrite";
