+++ src/components/Board/BoardDetail/BoardDetailStyle.jsx
