+++ src/components/Board/BoardFetchItems.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect, useRef } from "react";
 import { Link, useLocation } from "react-router-dom";
 import styled from "styled-components";
-import BoardItemBox from "../ThreePower/ThreePowerForm";
+import BoardItemBox from "./BoardStyle/ThreePowerForm";
 import axios from "axios";
-import BoardForm from "./BoardForm";
+import BoardForm from "./BoardStyle/BoardForm";
 
 export default function BoardFetchItems({ category = "", center }) {
   const { pathname } = useLocation();
