+++ src/components/Board/BoardStyle/BoardStyle.jsx
