+++ src/components/Board/NewWrite/BoardWriteStyle.jsx
