+++ src/components/Board/NewWrite/ThreeNewWrite.jsx
@@ -1,4 +1,4 @@
-import { useLocation, useNavigate } from "react-router";
+import { useNavigate } from "react-router";
 import Tiptap from "../../../utils/Editor/Tiptap";
 import { useState } from "react";
 import VideoIcon from "../../../assets/images/board_write_video_24.svg";
@@ -10,7 +10,7 @@ import {
   FileList,
   FileBtnDiv,
   FileBtn,
-} from "../BoardWriteStyle";
+} from "./BoardWriteStyle";
 
 export default function ThreeBoardPost() {
   let navigate = useNavigate();
@@ -85,7 +85,11 @@ export default function ThreeBoardPost() {
             <div style={{ display: "flex", flexDirection: "column" }}>
               <FileList>
                 <PostLabel>
-                  <img src={VideoIcon} style={{ marginRight: "5px" }} />
+                  <img
+                    src={VideoIcon}
+                    style={{ marginRight: "5px" }}
+                    alt="videoicon"
+                  />
                   벤치프레스 영상
                   <input
                     style={{ display: "none" }}
@@ -100,7 +104,11 @@ export default function ThreeBoardPost() {
               </FileList>
               <FileList>
                 <PostLabel>
-                  <img src={VideoIcon} style={{ marginRight: "5px" }} />
+                  <img
+                    src={VideoIcon}
+                    style={{ marginRight: "5px" }}
+                    alt="videoicon"
+                  />
                   데드리프트 영상
                   <input
                     onChange={changeDead}
@@ -115,7 +123,11 @@ export default function ThreeBoardPost() {
               </FileList>
               <FileList>
                 <PostLabel>
-                  <img src={VideoIcon} style={{ marginRight: "5px" }} />
+                  <img
+                    src={VideoIcon}
+                    style={{ marginRight: "5px" }}
+                    alt="videoicon"
+                  />
                   스쿼트 영상
                   <input
                     onChange={changeSqaut}
