+++ src/components/Board/ThreePower/ManagerThreePower.jsx
