+++ src/pages/ThreeBoardEditPost.jsx
@@ -1,47 +1,12 @@
 import { useLocation, useNavigate } from "react-router";
-import Tiptap from "../../../utils/Editor/Tiptap";
+import Tiptap from "../utils/Editor/Tiptap";
 import { useEffect, useState } from "react";
-import styled from "styled-components";
 import axios from "axios";
-import Nav from "../../Navbar";
-
-const PostWrapper = styled.div`
-  max-width: 1040px;
-  padding: 60px 16px;
-  width: 1024px;
-  margin: 0 auto;
-  color: #333;
-  grid-gap: 50px;
-  gap: 50px px;
-  position: relative;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  padding-top: 150px;
-`;
-
-const PostTitleTitle = styled.input`
-  box-sizing: border-box;
-  width: 100%;
-  height: 56px;
-  min-height: 56px;
-  line-height: 44px;
-  box-shadow: none;
-  padding-left: 16px;
-  padding-right: 52px;
-  border: 1px solid grey;
-  border-radius: 5px;
-  font-size: 16px;
-  font-weight: 500;
-  color: #333;
-`;
-
-const TitleLabel = styled.label`
-  display: inline-block;
-  margin-bottom: 5px;
-  color: #333;
-  font-weight: 500;
-`;
+import Nav from "../components/Navbar";
+import {
+  PostWrapper,
+  PostTitleTitle,
+} from "../components/Board/NewWrite/BoardWriteStyle";
 
 export default function ThreeBoardEditPost() {
   let navigate = useNavigate();
@@ -66,7 +31,7 @@ export default function ThreeBoardEditPost() {
       setLoading(false);
     };
     fetchData();
-  }, []);
+  }, [url]);
   const handleSubmit = (e) => {
     e.preventDefault();
     let benchFile = e.target.bench.files[0];
@@ -110,7 +75,6 @@ export default function ThreeBoardEditPost() {
         </div>
       ) : (
         <PostWrapper>
-          <TitleLabel Htmlfor="input-title">제목</TitleLabel>
           <form onSubmit={handleSubmit}>
             <PostTitleTitle
               id="input-title"
